feat(core): add catalog module for model metadata retrieval
- Introduced a new `catalog.ts` file to provide the `getModelMeta` function for retrieving model metadata from a catalog.
- Updated `package.json` to include the new catalog types in the exports.
- Enhanced the main index file to re-export the `getModelMeta` function for easier access.
- Updated documentation in the README files to include usage examples for the new metadata retrieval functionality.
