+++ packages/core/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./registry.js";
 export * from "./types.js";
+export * from "./catalog.js";
