+++ packages/helpers/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tokenlens/helpers",
-  "version": "1.0.0-beta.1",
+  "version": "1.0.0-beta.2",
   "private": false,
   "description": "Helpers for context windows, usage normalization, compaction, and cost estimation.",
   "type": "module",
