+++ packages/models/src/api.ts
@@ -1,18 +1,5 @@
-import type {
-  Model,
-  ModelCatalog,
-  ProviderInfo,
-  ProviderModel,
-} from "@tokenlens/core";
-import { toModelId } from "@tokenlens/core/id";
-
-export type GetModelMetaOptions = {
-  providers: ModelCatalog;
-  provider?: string;
-  model?: string;
-  models?: readonly string[];
-  id?: string;
-};
+import type { Model, ModelCatalog } from "@tokenlens/core";
+export { getModelMeta } from "@tokenlens/core";
 
 export type PickMap = Record<string, ReadonlyArray<string> | undefined>;
 
@@ -68,110 +55,12 @@ export function catalogFromModelArrays(
 }
 
 export function catalogFromProviders(
-  providers: ReadonlyArray<ProviderInfo>,
+  providers: ReadonlyArray<import("@tokenlens/core").ProviderInfo>,
 ): ModelCatalog {
   const out: ModelCatalog = {} as ModelCatalog;
   for (const p of providers) {
     if (!p?.id) continue;
-    out[p.id] = p as ProviderInfo;
-  }
-  return out;
-}
-
-/**
- * Get raw provider info or one/many model entries directly from the static
- * provider files generated in this package.
- *
- * - getModelMeta(provider) -> ProviderInfo | undefined
- * - getModelMeta(provider, modelId) -> ProviderModel | undefined
- * - getModelMeta(provider, [ids...]) -> Record<id, ProviderModel>
- */
-// Object overloads (preferred)
-export function getModelMeta(opts: {
-  providers: ModelCatalog;
-  provider: string;
-  model?: undefined;
-  models?: undefined;
-  id?: undefined;
-}): ProviderInfo | undefined;
-export function getModelMeta(opts: {
-  providers: ModelCatalog;
-  provider: string;
-  model: string;
-  models?: undefined;
-  id?: undefined;
-}): ProviderModel | undefined;
-export function getModelMeta(opts: {
-  providers: ModelCatalog;
-  provider: string;
-  models: readonly string[];
-  model?: undefined;
-  id?: undefined;
-}): Record<string, ProviderModel>;
-export function getModelMeta(opts: {
-  providers: ModelCatalog;
-  id: string;
-  provider?: string;
-  model?: string;
-  models?: undefined;
-}): ProviderModel | undefined;
-export function getModelMeta(
-  opts: GetModelMetaOptions,
-): ProviderInfo | ProviderModel | Record<string, ProviderModel> | undefined;
-// Positional overloads (back-compat)
-export function getModelMeta(
-  providers: ModelCatalog,
-  providerOrId: string,
-): ProviderInfo | ProviderModel | undefined;
-export function getModelMeta(
-  providers: ModelCatalog,
-  providerOrId: string,
-  modelId: string,
-): ProviderModel | undefined;
-export function getModelMeta(
-  providers: ModelCatalog,
-  providerOrId: string,
-  modelIds: readonly string[],
-): Record<string, ProviderModel>;
-export function getModelMeta(
-  a: GetModelMetaOptions | ModelCatalog,
-  b?: string,
-  c?: string | readonly string[],
-): ProviderInfo | ProviderModel | Record<string, ProviderModel> | undefined {
-  let opts: GetModelMetaOptions;
-  if (typeof a === "object" && "providers" in a && b === undefined) {
-    opts = a as GetModelMetaOptions;
-  } else {
-    opts = {
-      providers: a as ModelCatalog,
-      provider: b as string,
-      model: typeof c === "string" ? (c as string) : undefined,
-      models: Array.isArray(c) ? (c as readonly string[]) : undefined,
-    };
-  }
-
-  let provider = opts.provider;
-  let mm: string | readonly string[] | undefined = opts.models ?? opts.model;
-  if (!mm && opts.id) {
-    const canonical = toModelId(opts.id) ?? opts.id.replace("/", ":");
-    const idx = canonical.indexOf(":");
-    if (idx > 0) {
-      provider = canonical.slice(0, idx);
-      mm = canonical.slice(idx + 1);
-    }
-  }
-
-  if (!provider) return undefined;
-  const provInfo = opts.providers[provider] as ProviderInfo | undefined;
-  if (!provInfo) return undefined;
-  if (mm === undefined) return provInfo;
-  if (typeof mm === "string") {
-    return provInfo.models[mm] as ProviderModel | undefined;
-  }
-  const out: Record<string, ProviderModel> = {};
-  for (const id of mm) {
-    const hit = provInfo.models[id];
-    if (hit) out[id] = hit as ProviderModel;
+    out[p.id] = p as import("@tokenlens/core").ProviderInfo;
   }
   return out;
 }
