+++ packages/tokenlens/package.json
@@ -1,6 +1,6 @@
 {
   "name": "tokenlens",
-  "version": "1.3.0-canary.4",
+  "version": "1.3.0-canary.5",
   "description": "A lightweight registry of LLM model information, like name and context sizes, for building AI-powered apps.",
   "files": [
     "dist"
@@ -39,7 +39,7 @@
     "access": "public"
   },
   "scripts": {
-    "prebuild": "node scripts/gen-providers.mjs",
+    "prebuild": "pnpm --filter @tokenlens/core run build && pnpm --filter @tokenlens/helpers run build && pnpm --filter @tokenlens/fetch run build && pnpm --filter @tokenlens/models run build && node scripts/gen-providers.mjs",
     "build": "tsc",
     "typecheck": "tsc --noEmit",
     "test": "pnpm run build && vitest",
